+++ docs/CHANGELOG.rst
@@ -22,6 +22,7 @@ Added
 - Add CNVKit, LoFreq and GATK to ``resolwebio/dnaseq`` docker image
 - Add BaseSpace files download tool
 - Add process to import a file from BaseSpace
+- Add process to convert files to single-end reads
 
 Fixed
 -----